refactor shared actions and selectors to shared directory
