+++ src/applications/vaos/appointment-list/redux/reducer.js
@@ -2,31 +2,27 @@ import moment from 'moment';
 import set from 'platform/utilities/data/set';
 
 import {
-  FETCH_FUTURE_APPOINTMENTS,
-  FETCH_FUTURE_APPOINTMENTS_SUCCEEDED,
-  FETCH_FUTURE_APPOINTMENTS_FAILED,
-  FETCH_PENDING_APPOINTMENTS,
-  FETCH_PENDING_APPOINTMENTS_SUCCEEDED,
-  FETCH_PENDING_APPOINTMENTS_FAILED,
-  FETCH_PAST_APPOINTMENTS,
-  FETCH_PAST_APPOINTMENTS_SUCCEEDED,
-  FETCH_PAST_APPOINTMENTS_FAILED,
-  FETCH_REQUEST_DETAILS,
-  FETCH_REQUEST_DETAILS_SUCCEEDED,
   CANCEL_APPOINTMENT,
+  CANCEL_APPOINTMENT_CLOSED,
   CANCEL_APPOINTMENT_CONFIRMED,
   CANCEL_APPOINTMENT_CONFIRMED_FAILED,
   CANCEL_APPOINTMENT_CONFIRMED_SUCCEEDED,
-  CANCEL_APPOINTMENT_CLOSED,
-  FETCH_FACILITY_LIST_DATA_SUCCEEDED,
   FETCH_CONFIRMED_DETAILS,
-  FETCH_CONFIRMED_DETAILS_SUCCEEDED,
   FETCH_CONFIRMED_DETAILS_FAILED,
-  FETCH_REQUEST_DETAILS_FAILED,
+  FETCH_CONFIRMED_DETAILS_SUCCEEDED,
+  FETCH_FACILITY_SETTINGS,
   FETCH_FACILITY_SETTINGS_FAILED,
   FETCH_FACILITY_SETTINGS_SUCCEEDED,
-  FETCH_FACILITY_SETTINGS,
+  FETCH_FUTURE_APPOINTMENTS,
+  FETCH_FUTURE_APPOINTMENTS_FAILED,
+  FETCH_FUTURE_APPOINTMENTS_SUCCEEDED,
+  FETCH_PAST_APPOINTMENTS,
+  FETCH_PAST_APPOINTMENTS_FAILED,
+  FETCH_PAST_APPOINTMENTS_SUCCEEDED,
   FETCH_PROVIDER_SUCCEEDED,
+  FETCH_REQUEST_DETAILS,
+  FETCH_REQUEST_DETAILS_FAILED,
+  FETCH_REQUEST_DETAILS_SUCCEEDED,
 } from './actions';
 
 import {
@@ -34,7 +30,13 @@ import {
   VACCINE_FORM_SUBMIT_SUCCEEDED,
 } from '../../redux/sitewide';
 
-import { FETCH_STATUS, APPOINTMENT_STATUS } from '../../utils/constants';
+import {
+  FETCH_FACILITY_LIST_DATA_SUCCEEDED,
+  FETCH_PENDING_APPOINTMENTS,
+  FETCH_PENDING_APPOINTMENTS_FAILED,
+  FETCH_PENDING_APPOINTMENTS_SUCCEEDED,
+} from '../../redux/actions';
+import { APPOINTMENT_STATUS, FETCH_STATUS } from '../../utils/constants';
 
 const initialState = {
   pending: null,
